source code/python/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_ipyw_jlab_nb_ext_conf=0.1.0=py36he6757f0_0
alabaster=0.7.10=py36hcd07829_0
anaconda-client=1.6.9=py36_0
anaconda-navigator=1.7.0=py36_0
anaconda-project=0.8.2=py36hfad2e28_0
asn1crypto=0.24.0=py36_0
astroid=1.6.1=py36_0
astropy=2.0.3=py36hfa6e2cd_0
attrs=17.4.0=py36_0
babel=2.5.3=py36_0
backports=1.0=py36h81696a8_1
backports.shutil_get_terminal_size=1.0.0=py36h79ab834_2
backports.weakref=1.0rc1=py36_0
beautifulsoup4=4.6.0=py36hd4cc5e8_1
bitarray=0.8.1=py36hfa6e2cd_1
bkcharts=0.2=py36h7e685f7_0
blaze=0.11.3=py36h8a29ca5_0
bleach=1.5.0=py36_0
bokeh=0.12.13=py36h047fa9f_0
boto=2.48.0=py36h1a776d2_1
bottleneck=1.2.1=py36hd119dfa_0
bzip2=1.0.6=hbe05fcf_4
ca-certificates=2018.4.16=0
certifi=2018.4.16=py36_0
cffi=1.11.4=py36hfa6e2cd_0
chardet=3.0.4=py36h420ce6e_1
click=6.7=py36hec8c647_0
cloudpickle=0.5.2=py36_1
clyent=1.2.2=py36hb10d595_1
colorama=0.3.9=py36h029ae33_0
comtypes=1.1.4=py36_0
conda=4.4.10=py36_0
conda-build=3.4.1=py36_0
conda-env=2.6.0=h36134e3_1
conda-verify=2.0.0=py36h065de53_0
console_shortcut=0.1.1=h6bb2dd7_3
contextlib2=0.5.5=py36he5d52c0_0
cryptography=2.1.4=py36he1d7878_0
curl=7.58.0=h7602738_0
cycler=0.10.0=py36h009560c_0
cython=0.27.3=py36h22f4c84_0
cytoolz=0.9.0=py36hfa6e2cd_0
dask=0.16.1=py36_0
dask-core=0.16.1=py36_0
datashape=0.5.4=py36h5770b85_0
decorator=4.2.1=py36_0
distributed=1.20.2=py36_0
docutils=0.14=py36h6012d8f_0
entrypoints=0.2.3=py36hfd66bb0_2
et_xmlfile=1.0.1=py36h3d2d736_0
fastcache=1.0.2=py36hfa6e2cd_2
fbprophet=0.2.1=py36_vc14_0
filelock=2.0.13=py36h20000bf_0
flask=0.12.2=py36h98b5e8f_0
flask-cors=3.0.3=py36h8a3855d_0
freetype=2.8.1=vc14_0
get_terminal_size=1.0.0=h38e98db_0
gevent=1.2.2=py36h342a76c_0
glob2=0.6=py36hdf76b57_0
greenlet=0.4.12=py36ha00ad21_0
gurobi=8.0.0=py36hea7dcea_0
h5py=2.7.1=py36he54a1c3_0
hdf5=1.10.1=vc14_2
heapdict=1.0.0=py36_2
html5lib=0.9999999=py36_0
icc_rt=2017.0.4=h97af966_0
icu=58.2=vc14_0
idna=2.6=py36h148d497_1
imageio=2.2.0=py36had6c2d2_0
imagesize=0.7.1=py36he29f638_0
intel-openmp=2018.0.0=hd92c6cd_8
ipykernel=4.8.0=py36_0
ipython=6.2.1=py36h9cf0123_1
ipython_genutils=0.2.0=py36h3c5d0ee_0
ipywidgets=7.1.1=py36_0
isort=4.2.15=py36h6198cc5_0
itsdangerous=0.24=py36hb6c5a24_1
jdcal=1.3=py36h64a5255_0
jedi=0.11.1=py36_0
jinja2=2.10=py36h292fed1_0
jpeg=9b=vc14_2
jsonschema=2.6.0=py36h7636477_0
jupyter=1.0.0=py36_4
jupyter_client=5.2.2=py36_0
jupyter_console=5.2.0=py36h6d89b47_1
jupyter_core=4.4.0=py36h56e9d50_0
jupyterlab=0.31.4=py36_0
jupyterlab_launcher=0.10.2=py36_0
keras=2.1.5=py36_0
lazy-object-proxy=1.3.1=py36hd1c21d2_0
libcurl=7.58.0=h7602738_0
libiconv=1.15=h1df5818_7
libpng=1.6.34=vc14_0
libprotobuf=3.5.2=vc14_0
libssh2=1.8.0=hd619d38_4
libtiff=4.0.9=vc14_0
libxml2=2.9.8=vc14_0
libxslt=1.1.32=vc14_0
llvmlite=0.21.0=py36he0b0552_0
locket=0.2.0=py36hfed976d_1
lxml=4.1.1=py36hef2cd61_1
lzo=2.10=h6df0209_2
markdown=2.6.11=py36_0
markupsafe=1.0=py36h0e26971_1
matplotlib=2.1.2=py36h016c42a_0
mccabe=0.6.1=py36hb41005a_1
menuinst=1.4.11=py36hfa6e2cd_0
mistune=0.8.3=py36_0
mkl=2018.0.1=h2108138_4
mkl-service=1.1.2=py36h57e144c_4
mpmath=1.0.0=py36hacc8adf_2
msgpack-python=0.5.1=py36he980bc4_0
multipledispatch=0.4.9=py36he44c36e_0
navigator-updater=0.1.0=py36h8a7b86b_0
nbconvert=5.3.1=py36h8dc0fde_0
nbformat=4.4.0=py36h3a5bc1b_0
networkx=2.1=py36_0
nltk=3.2.5=py36h76d52bb_0
nose=1.3.7=py36h1c3779e_2
notebook=5.4.0=py36_0
numba=0.36.2=np114py36h12cb543_0
numexpr=2.6.4=py36h30784b8_0
numpy=1.14.0=py36h4a99626_1
numpydoc=0.7.0=py36ha25429e_0
odo=0.5.1=py36h7560279_0
olefile=0.45.1=py36_0
openpyxl=2.4.10=py36_0
openssl=1.0.2o=vc14_0
packaging=16.8=py36ha0986f6_1
pandas=0.22.0=py36h6538335_0
pandoc=1.19.2.1=hb2460c7_1
pandocfilters=1.4.2=py36h3ef6317_1
parso=0.1.1=py36hae3edee_0
partd=0.3.8=py36hc8e763b_0
path.py=10.5=py36h2b94a8f_0
pathlib2=2.3.0=py36h7bfb78b_0
patsy=0.5.0=py36_0
pep8=1.7.1=py36_0
pickleshare=0.7.4=py36h9de030f_0
pillow=5.0.0=py36h0738816_0
pip=9.0.1=py36h226ae91_4
pkginfo=1.4.1=py36hb0f9cfa_1
pluggy=0.6.0=py36hc7daf1e_0
ply=3.10=py36h1211beb_0
prompt_toolkit=1.0.15=py36h60b8f86_0
protobuf=3.5.2=py36_vc14_0
psutil=5.4.3=py36hfa6e2cd_0
py=1.5.2=py36hbcfbabc_0
pycodestyle=2.3.1=py36h7cc55cd_0
pycosat=0.6.3=py36h413d8a4_0
pycparser=2.18=py36hd053e01_1
pycrypto=2.6.1=py36hfa6e2cd_7
pycurl=7.43.0.1=py36h74b6da3_0
pyflakes=1.6.0=py36h0b975d6_0
pygments=2.2.0=py36hb010967_0
pylint=1.8.2=py36_0
pyodbc=4.0.22=py36h6538335_0
pyopenssl=17.5.0=py36h5b7d817_0
pyparsing=2.2.0=py36h785a196_1
pyqt=5.6.0=py36hb5ed885_5
pysocks=1.6.7=py36h698d350_1
pystan=2.17.1.0=py36_vc14_0
pytables=3.4.2=py36h71138e3_2
pytest=3.3.2=py36_0
python=3.6.4=h6538335_1
python-dateutil=2.6.1=py36h509ddcb_1
pytz=2017.3=py36h1d3fa6b_0
pywavelets=0.5.2=py36hc649158_0
pywin32=222=py36hfa6e2cd_0
pywinpty=0.5=py36h6538335_1
pyyaml=3.12=py36h1d1928f_1
pyzmq=16.0.3=py36he714bf5_0
qt=5.6.2=vc14_1
qtawesome=0.4.4=py36h5aa48f6_0
qtconsole=4.3.1=py36h99a29a9_0
qtpy=1.3.1=py36hb8717c5_0
requests=2.18.4=py36h4371aae_1
rope=0.10.7=py36had63a69_0
ruamel_yaml=0.15.35=py36hfa6e2cd_1
scikit-image=0.13.1=py36hfa6e2cd_1
scikit-learn=0.19.1=py36h53aea1b_0
scipy=1.0.0=py36h1260518_0
seaborn=0.8.1=py36h9b69545_0
send2trash=1.4.2=py36_0
setuptools=38.4.0=py36_0
simplegeneric=0.8.1=py36_2
singledispatch=3.4.0.3=py36h17d0c80_0
sip=4.18.1=py36h9c25514_2
six=1.11.0=py36h4db2310_1
snowballstemmer=1.2.1=py36h763602f_0
sortedcollections=0.5.3=py36hbefa0ab_0
sortedcontainers=1.5.9=py36_0
sphinx=1.6.6=py36_0
sphinxcontrib=1.0=py36hbbac3d2_1
sphinxcontrib-websupport=1.0.1=py36hb5e5916_1
spyder=3.2.6=py36_0
sqlalchemy=1.2.1=py36hfa6e2cd_0
sqlite=3.22.0=vc14_0
statsmodels=0.8.0=py36h6189b4c_0
sympy=1.1.1=py36h96708e0_0
tblib=1.3.2=py36h30f5020_0
tensorflow=1.2.1=py36_0
terminado=0.8.1=py36_1
testpath=0.3.1=py36h2698cfe_0
tk=8.6.7=vc14_0
toolz=0.9.0=py36_0
tornado=4.5.3=py36_0
traitlets=4.3.2=py36h096827d_0
typing=3.6.2=py36hb035bda_0
unicodecsv=0.14.1=py36h6450c06_0
urllib3=1.22=py36h276f60a_0
vc=14=h0510ff6_3
vs2015_runtime=14.0.25420=0
wcwidth=0.1.7=py36h3d5aa90_0
webencodings=0.5.1=py36h67c50ae_1
werkzeug=0.14.1=py36_0
wheel=0.30.0=py36h6c3ec14_1
widgetsnbextension=3.1.0=py36_0
win_inet_pton=1.0.1=py36he67d7fd_1
win_unicode_console=0.5=py36hcdbd4b5_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
wrapt=1.10.11=py36he5f5981_0
xlrd=1.1.0=py36h1cb58dc_1
xlsxwriter=1.0.2=py36hf723b7d_0
xlwings=0.11.5=py36_0
xlwt=1.3.0=py36h1a4751e_0
yaml=0.1.7=vc14_0
zict=0.1.3=py36h2d8e73e_0
zlib=1.2.11=vc14_0
